+++ app/(root)/freeBoard/[articleId]/_components/EditDeleteMenu.jsx
