+++ app/(root)/layout.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import Header from "./_components/Header";
 import { basicFont } from "@/assets/fonts";
 import Footer from "./_components/Footer";
+
 function Rootlayout({ children }) {
   return (
     <div
