+++ app/globals.css
@@ -3,7 +3,6 @@
 @tailwind utilities;
 
 body {
-  font-family: "Pretendard Variable", sans-serif;
   margin: 0;
   padding: 0;
 }
