+++ assets/fonts/index.js
@@ -1,5 +1,5 @@
 import localFont from "next/font/local";
 
 export const basicFont = localFont({
-  src: "PretendardVariable.woff2",
+  src: "./PretendardVariable.woff2",
 });
